qvterm: use glyphs over individual chars
This is a huge performance bump that at least puts sff in the range of
other terminals.  I knew that writing cell by cell wasn't great and that
I needed to buffer but all previous attempts failed at layout.  However,
using QRawFont and QGlyphRun is working and is probably the right way to
interface with libvterm.

With this, sff running `top -d1` goes from taking 1-3% cpu to being
unnoticable.  `time dmesg` is much closer to pangoterm, st and urxvt.