qvterm: smaller update footprint for movecursor
Qt is going to stack our updates anyways, so there's no reason to try to
be smart.  Especially when the previous attempt at doing so was broken
and sometimes leaving a cursor behind.

The severity of this has gone back and forth for a while in QVTerm but
just prior to this commit opening a file in vim with syntax highlighting
and hitting <enter> over and over in the middle of the file in insert
mode was leaving cursors every time.