qvterm: fix col/row calculation
This was leaving artifacts from the previous redraw if we hit the edge
of the height or width of the window.  Likely it was working around
previous bugs that have since been fixed.

To reproduce the issue fixed type `echo <x>` with x being over the
width of the terminal, then without hitting enter, kill the line.  It
was leaving character(s) on the right hand side of the window.